ref(demo-spartan): Inline BrnFieldA11yService bridge factory
Replace the wrapper-level bridge class with an inline BrnFieldA11yService
factory in spartan-form-field, keeping the aria-describedby interop
contract while reducing indirection.

Update demo-spartan docs and related smoke/e2e spec commentary to match
the new provider shape, and remove the obsolete bridge file.

Refs #48
